FPGA: compile - remove +/- in path (#1404)
